+++ artiq/coredevice/__init__.py
@@ -1,9 +1,3 @@
-from artiq.coredevice import exceptions, dds, spi2
 from artiq.coredevice.exceptions import (RTIOUnderflow, RTIOOverflow)
-from artiq.coredevice.dds import (PHASE_MODE_CONTINUOUS, PHASE_MODE_ABSOLUTE,
-                                  PHASE_MODE_TRACKING)
 
-__all__ = []
-__all__ += ["RTIOUnderflow", "RTIOOverflow"]
-__all__ += ["PHASE_MODE_CONTINUOUS", "PHASE_MODE_ABSOLUTE",
-            "PHASE_MODE_TRACKING"]
+__all__ = ["RTIOUnderflow", "RTIOOverflow"]