+++ artiq/examples/kc705_nist_clock/repository/photon_histogram.py
@@ -6,7 +6,6 @@ class PhotonHistogram(EnvExperiment):
 
     def build(self):
         self.setattr_device("core")
-        self.setattr_device("core_dds")
         self.setattr_device("bd_dds")
         self.setattr_device("bd_sw")
         self.setattr_device("bdd_dds")
@@ -22,9 +21,10 @@ class PhotonHistogram(EnvExperiment):
 
     @kernel
     def program_cooling(self):
-        with self.core_dds.batch:
-            self.bd_dds.set(200*MHz)
-            self.bdd_dds.set(300*MHz)
+        delay_mu(-self.bd_dds.set_duration_mu)
+        self.bd_dds.set(200*MHz)
+        delay_mu(self.bd_dds.set_duration_mu)
+        self.bdd_dds.set(300*MHz)
 
     @kernel
     def cool_detect(self):